Made sure oq export look in the database first [skip hazardlib][demos]
